+++ jersey/src/main/java/AppConfig.java
@@ -1,5 +1,3 @@
-import service.ServletContainer;
-
 import javax.ws.rs.core.Application;
 import java.util.HashSet;
 import java.util.Set;
